@@ -20,7 +20,7 @@
         }
     ],
     "require": {
-        "php": "^5.6 || ^7.0",
+        "php": "^7.1",
         "doctrine/common": "^2.3",
         "sonata-project/cache": "^1.0 || ^2.0",
         "sonata-project/core-bundle": "^3.4",
@@ -34,7 +34,6 @@
         "symfony/http-foundation": "^2.8 || ^3.2 || ^4.0",
         "symfony/http-kernel": "^2.8 || ^3.2 || ^4.0",
         "symfony/options-resolver": "^2.8 || ^3.2 || ^4.0",
-        "symfony/templating": "^2.8 || ^3.2 || ^4.0",
         "symfony/twig-bundle": "^2.8 || ^3.2 || ^4.0",
         "twig/twig": "^1.34 || ^2.0"
     },
@@ -44,7 +43,7 @@
     "require-dev": {
         "jms/di-extra-bundle": "^1.7",
         "knplabs/knp-menu-bundle": "^2.0",
-        "sonata-project/admin-bundle": "^3.22",
+        "sonata-project/admin-bundle": "^3.22 || ^4.0@dev",
         "symfony/debug": "^2.8 || ^3.2 || ^4.0",
         "symfony/phpunit-bridge": "^4.0",
         "symfony/stopwatch": "^2.8 || ^3.2 || ^4.0"
@@ -59,16 +58,13 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.x-dev"
+            "dev-master": "4.x-dev"
         }
     },
     "autoload": {
         "psr-4": {
             "Sonata\\BlockBundle\\": "src/"
-        },
-        "files": [
-            "src/Resources/stubs/symfony2.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
