@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -59,7 +61,7 @@ class ServiceLoader implements BlockLoaderInterface
         $block->setEnabled(true);
         $block->setCreatedAt(new \DateTime());
         $block->setUpdatedAt(new \DateTime());
-        $block->setSettings(isset($configuration['settings']) ? $configuration['settings'] : []);
+        $block->setSettings($configuration['settings'] ?? []);
 
         return $block;
     }
