@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -33,7 +35,7 @@ class ContainerBlockService extends AbstractAdminBlockService
     /**
      * {@inheritdoc}
      */
-    public function buildEditForm(FormMapper $formMapper, BlockInterface $block)
+    public function buildEditForm(FormMapper $formMapper, BlockInterface $block): void
     {
         $formMapper->add('enabled');
 
@@ -80,7 +82,7 @@ class ContainerBlockService extends AbstractAdminBlockService
     /**
      * {@inheritdoc}
      */
-    public function configureSettings(OptionsResolver $resolver)
+    public function configureSettings(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'code' => '',
@@ -116,8 +118,8 @@ class ContainerBlockService extends AbstractAdminBlockService
 
         $segments = explode($key, $layout);
         $decorator = [
-            'pre' => isset($segments[0]) ? $segments[0] : '',
-            'post' => isset($segments[1]) ? $segments[1] : '',
+            'pre' => $segments[0] ?? '',
+            'post' => $segments[1] ?? '',
         ];
 
         return $decorator;
