@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -33,7 +35,7 @@ class Block implements MetadataProcessorInterface
     /**
      * @param ClassMetadata $metadata
      */
-    public function processMetadata(ClassMetadata $metadata)
+    public function processMetadata(ClassMetadata $metadata): void
     {
         if (!empty($this->id)) {
             $metadata->id = $this->id;
