@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,7 +18,6 @@ use Psr\Log\LoggerInterface;
 use Sonata\BlockBundle\Model\BlockInterface;
 use Symfony\Component\OptionsResolver\Exception\ExceptionInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 class BlockContextManager implements BlockContextManagerInterface
 {
@@ -51,14 +52,6 @@ class BlockContextManager implements BlockContextManagerInterface
     protected $logger;
 
     /**
-     * Used for deprecation check on {@link resolve} method.
-     * To be removed in 4.0 with BC system.
-     *
-     * @var array
-     */
-    private $reflectionCache;
-
-    /**
      * @param BlockLoaderInterface         $blockLoader
      * @param BlockServiceManagerInterface $blockService
      * @param array                        $cacheBlocks
@@ -71,13 +64,12 @@ class BlockContextManager implements BlockContextManagerInterface
         $this->blockService = $blockService;
         $this->cacheBlocks = $cacheBlocks;
         $this->logger = $logger;
-        $this->reflectionCache = [];
     }
 
     /**
      * {@inheritdoc}
      */
-    public function addSettingsByType($type, array $settings, $replace = false)
+    public function addSettingsByType($type, array $settings, $replace = false): void
     {
         $typeSettings = isset($this->settingsByType[$type]) ? $this->settingsByType[$type] : [];
         if ($replace) {
@@ -90,7 +82,7 @@ class BlockContextManager implements BlockContextManagerInterface
     /**
      * {@inheritdoc}
      */
-    public function addSettingsByClass($class, array $settings, $replace = false)
+    public function addSettingsByClass($class, array $settings, $replace = false): void
     {
         $classSettings = isset($this->settingsByClass[$class]) ? $this->settingsByClass[$class] : [];
         if ($replace) {
@@ -155,28 +147,7 @@ class BlockContextManager implements BlockContextManagerInterface
         return $blockContext;
     }
 
-    /**
-     * NEXT_MAJOR: remove this method.
-     *
-     * @param OptionsResolverInterface $optionsResolver
-     * @param BlockInterface           $block
-     *
-     * @deprecated since version 2.3, to be renamed in 4.0.
-     *             Use the method configureSettings instead
-     */
-    protected function setDefaultSettings(OptionsResolverInterface $optionsResolver, BlockInterface $block)
-    {
-        if (__CLASS__ !== get_called_class()) {
-            @trigger_error(
-                'The '.__METHOD__.' is deprecated since version 2.3, to be renamed in 4.0.'
-                .' Use '.__CLASS__.'::configureSettings instead.',
-                E_USER_DEPRECATED
-            );
-        }
-        $this->configureSettings($optionsResolver, $block);
-    }
-
-    protected function configureSettings(OptionsResolver $optionsResolver, BlockInterface $block)
+    protected function configureSettings(OptionsResolver $optionsResolver, BlockInterface $block): void
     {
         // defaults for all blocks
         $optionsResolver->setDefaults([
@@ -188,12 +159,12 @@ class BlockContextManager implements BlockContextManagerInterface
         ]);
 
         $optionsResolver
-            ->addAllowedTypes('use_cache', 'bool')
-            ->addAllowedTypes('extra_cache_keys', 'array')
-            ->addAllowedTypes('attr', 'array')
-            ->addAllowedTypes('ttl', 'int')
-            ->addAllowedTypes('template', ['string', 'bool'])
-        ;
+                ->addAllowedTypes('use_cache', 'bool')
+                ->addAllowedTypes('extra_cache_keys', 'array')
+                ->addAllowedTypes('attr', 'array')
+                ->addAllowedTypes('ttl', 'int')
+                ->addAllowedTypes('template', ['string', 'bool'])
+            ;
 
         // add type and class settings for block
         $class = ClassUtils::getClass($block);
@@ -209,7 +180,7 @@ class BlockContextManager implements BlockContextManagerInterface
      * @param BlockContextInterface $blockContext
      * @param array                 $settings
      */
-    protected function setDefaultExtraCacheKeys(BlockContextInterface $blockContext, array $settings)
+    protected function setDefaultExtraCacheKeys(BlockContextInterface $blockContext, array $settings): void
     {
         if (!$blockContext->getSetting('use_cache') || $blockContext->getSetting('ttl') <= 0) {
             return;
@@ -251,7 +222,7 @@ class BlockContextManager implements BlockContextManagerInterface
      */
     private function resolve(BlockInterface $block, $settings)
     {
-        $optionsResolver = new \Sonata\BlockBundle\Util\OptionsResolver();
+        $optionsResolver = new OptionsResolver();
 
         $this->configureSettings($optionsResolver, $block);
 
