@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -141,7 +143,9 @@ class BlockRenderer implements BlockRendererInterface
     {
         // a response exists, use it
         if ($this->lastResponse && $this->lastResponse->isCacheable()) {
-            $response->setTtl($this->lastResponse->getTtl());
+            if($this->lastResponse->getTtl()<$response->getTtl()) {
+                $response->setTtl($this->lastResponse->getTtl());
+            }
             $response->setPublic();
         } elseif ($this->lastResponse) { // not cacheable
             $response->setPrivate();
