@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -13,7 +15,7 @@ namespace Sonata\BlockBundle\Block;
 
 use Sonata\BlockBundle\Model\BlockInterface;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Interface BlockServiceInterface.
@@ -36,19 +38,6 @@ interface BlockServiceInterface
      * @return string
      */
     public function getName();
-
-    /**
-     * Define the default options for the block.
-     *
-     * NEXT_MAJOR: rename this method.
-     *
-     * @param OptionsResolverInterface $resolver
-     *
-     * @deprecated since version 2.3, to be renamed in 4.0.
-     *             Use the method configureSettings instead.
-     *             This method will be added to the BlockServiceInterface with SonataBlockBundle 4.0
-     */
-    public function setDefaultSettings(OptionsResolverInterface $resolver);
 
     /**
      * @param BlockInterface $block
@@ -75,4 +64,11 @@ interface BlockServiceInterface
      * @return array
      */
     public function getCacheKeys(BlockInterface $block);
+
+    /**
+     * Define the default options for the block.
+     *
+     * @param OptionsResolver $resolver
+     */
+    public function configureSettings(OptionsResolver $resolver);
 }
