@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -79,7 +81,7 @@ class BlockServiceManager implements BlockServiceManagerInterface
     /**
      * {@inheritdoc}
      */
-    public function add($name, $service, $contexts = [])
+    public function add($name, $service, $contexts = []): void
     {
         $this->services[$name] = $service;
 
@@ -92,19 +94,6 @@ class BlockServiceManager implements BlockServiceManagerInterface
         }
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function setServices(array $blockServices)
-    {
-        foreach ($blockServices as $name => $service) {
-            $this->add($name, $service);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getServices()
     {
         foreach ($this->services as $name => $id) {
@@ -141,29 +130,11 @@ class BlockServiceManager implements BlockServiceManagerInterface
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function getLoadedServices()
-    {
-        $services = [];
-
-        foreach ($this->services as $service) {
-            if (!$service instanceof BlockServiceInterface) {
-                continue;
-            }
-
-            $services[] = $service;
-        }
-
-        return $services;
-    }
-
-    /**
      * @todo: this function should be remove into a proper statefull object
      *
      * {@inheritdoc}
      */
-    public function validate(ErrorElement $errorElement, BlockInterface $block)
+    public function validate(ErrorElement $errorElement, BlockInterface $block): void
     {
         if (!$block->getId() && !$block->getType()) {
             return;
