@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,32 +18,28 @@ use Sonata\BlockBundle\Block\BlockContextInterface;
 use Sonata\BlockBundle\Model\BlockInterface;
 use Sonata\CoreBundle\Validator\ErrorElement;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class EmptyBlockService extends AbstractBlockService
 {
     // NEXT_MAJOR: Remove this method
 
-    /**
-     * {@inheritdoc}
-     */
-    public function buildEditForm(FormMapper $form, BlockInterface $block)
+    public function configureSettings(OptionsResolver $resolver): void
+    {
+    }
+
+    public function buildEditForm(FormMapper $form, BlockInterface $block): void
     {
         throw new \RuntimeException('Not used, this block renders an empty result if no block document can be found');
     }
 
     // NEXT_MAJOR: Remove this method
 
-    /**
-     * {@inheritdoc}
-     */
-    public function validateBlock(ErrorElement $errorElement, BlockInterface $block)
+    public function validateBlock(ErrorElement $errorElement, BlockInterface $block): void
     {
         throw new \RuntimeException('Not used, this block renders an empty result if no block document can be found');
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function execute(BlockContextInterface $blockContext, Response $response = null)
     {
         return new Response();
