@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -25,7 +27,7 @@ class HttpCacheHandler implements HttpCacheHandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function alterResponse(Response $response)
+    public function alterResponse(Response $response): void
     {
         if (!$response->isCacheable()) {
             // the controller flags the response as private so we keep it private!
@@ -47,7 +49,7 @@ class HttpCacheHandler implements HttpCacheHandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function updateMetadata(Response $response, BlockContextInterface $blockContext = null)
+    public function updateMetadata(Response $response, BlockContextInterface $blockContext = null): void
     {
         if (null === $this->currentTtl) {
             $this->currentTtl = $response->getTtl();
@@ -61,7 +63,7 @@ class HttpCacheHandler implements HttpCacheHandlerInterface
     /**
      * {@inheritdoc}
      */
-    public function onKernelResponse(FilterResponseEvent $event)
+    public function onKernelResponse(FilterResponseEvent $event): void
     {
         $this->alterResponse($event->getResponse());
     }
