@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -25,7 +27,7 @@ class TweakCompilerPass implements CompilerPassInterface
     /**
      * {@inheritdoc}
      */
-    public function process(ContainerBuilder $container)
+    public function process(ContainerBuilder $container): void
     {
         $manager = $container->getDefinition('sonata.block.manager');
         $registry = $container->getDefinition('sonata.block.menu.registry');
@@ -36,25 +38,12 @@ class TweakCompilerPass implements CompilerPassInterface
             $definition = $container->getDefinition($id);
             $definition->setPublic(true);
 
-            $arguments = $definition->getArguments();
-
             // Replace empty block id with service id
-            if (empty($arguments) || 0 == strlen($arguments[0])) {
-                // NEXT_MAJOR: Remove the condition when Symfony 2.8 support will be dropped.
-                if (method_exists($definition, 'setArgument')) {
-                    $definition->setArgument(0, $id);
-                } else {
-                    $definition->replaceArgument(0, $id);
-                }
-            } elseif ($id != $arguments[0] && 0 !== strpos(
-                $container->getParameterBag()->resolveValue($definition->getClass()),
-                'Sonata\\BlockBundle\\Block\\Service\\'
-            )) {
-                // NEXT_MAJOR: Remove deprecation notice
-                @trigger_error(
-                    sprintf('Using service id %s different from block id %s is deprecated since 3.3 and will be removed in 4.0.', $id, $arguments[0]),
-                    E_USER_DEPRECATED
-                );
+            // NEXT_MAJOR: Remove the condition when Symfony 2.8 support will be dropped.
+            if (method_exists($definition, 'setArgument')) {
+                $definition->setArgument(0, $id);
+            } else {
+                $definition->replaceArgument(0, $id);
             }
 
             $manager->addMethodCall('add', [$id, $id, isset($parameters[$id]) ? $parameters[$id]['contexts'] : []]);
@@ -84,7 +73,7 @@ class TweakCompilerPass implements CompilerPassInterface
      *
      * @param ContainerBuilder $container
      */
-    public function applyContext(ContainerBuilder $container)
+    public function applyContext(ContainerBuilder $container): void
     {
         $definition = $container->findDefinition('sonata.block.context_manager');
 
