@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -48,7 +50,7 @@ class SonataBlockExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $bundles = $container->getParameter('kernel.bundles');
 
@@ -94,7 +96,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureBlockContainers(ContainerBuilder $container, array $config)
+    public function configureBlockContainers(ContainerBuilder $container, array $config): void
     {
         $container->setParameter('sonata.block.container.types', $config['container']['types']);
 
@@ -104,7 +106,7 @@ class SonataBlockExtension extends Extension
     /**
      * @param array $config
      */
-    public function fixConfigurationDeprecation(array &$config)
+    public function fixConfigurationDeprecation(array &$config): void
     {
         if (count(array_diff($config['profiler']['container_types'], $config['container']['types']))) {
             $config['container']['types'] = array_merge($config['profiler']['container_types'], $config['container']['types']);
@@ -115,7 +117,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureMenus(ContainerBuilder $container, array $config)
+    public function configureMenus(ContainerBuilder $container, array $config): void
     {
         $bundles = $container->getParameter('kernel.bundles');
         if (!isset($bundles['KnpMenuBundle'])) {
@@ -131,7 +133,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureContext(ContainerBuilder $container, array $config)
+    public function configureContext(ContainerBuilder $container, array $config): void
     {
         $container->setParameter($this->getAlias().'.blocks', $config['blocks']);
         $container->setParameter($this->getAlias().'.blocks_by_class', $config['blocks_by_class']);
@@ -143,7 +145,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureCache(ContainerBuilder $container, array $config)
+    public function configureCache(ContainerBuilder $container, array $config): void
     {
         $container->setAlias('sonata.block.cache.handler', $config['http_cache']['handler']);
 
@@ -167,7 +169,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureLoaderChain(ContainerBuilder $container, array $config)
+    public function configureLoaderChain(ContainerBuilder $container, array $config): void
     {
         $types = [];
         foreach ($config['blocks'] as $service => $settings) {
@@ -181,7 +183,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container
      * @param array            $config
      */
-    public function configureForm(ContainerBuilder $container, array $config)
+    public function configureForm(ContainerBuilder $container, array $config): void
     {
         $defaults = $config['default_contexts'];
 
@@ -208,7 +210,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container Container
      * @param array            $config    Configuration
      */
-    public function configureProfiler(ContainerBuilder $container, array $config)
+    public function configureProfiler(ContainerBuilder $container, array $config): void
     {
         $container->setAlias('sonata.block.renderer', 'sonata.block.renderer.default');
 
@@ -232,7 +234,7 @@ class SonataBlockExtension extends Extension
      * @param ContainerBuilder $container Container builder
      * @param array            $config    An array of configuration
      */
-    public function configureException(ContainerBuilder $container, array $config)
+    public function configureException(ContainerBuilder $container, array $config): void
     {
         // retrieve available filters
         $filters = [];
@@ -274,7 +276,7 @@ class SonataBlockExtension extends Extension
     /**
      * Add class to compile.
      */
-    public function configureClassesToCompile()
+    public function configureClassesToCompile(): void
     {
         $this->addClassesToCompile([
             'Sonata\\BlockBundle\\Block\\BlockLoaderChain',
