@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -37,7 +39,7 @@ class BlockEvent extends Event
     /**
      * @param BlockInterface $block
      */
-    public function addBlock(BlockInterface $block)
+    public function addBlock(BlockInterface $block): void
     {
         $this->blocks[] = $block;
     }
