@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -31,7 +33,7 @@ class TextBlockListener
     /**
      * @param BlockEvent $event
      */
-    public function onBlock(BlockEvent $event)
+    public function onBlock(BlockEvent $event): void
     {
         $content = 'This block is coming from inline event from the template';
         if ($event->getSetting('admin') instanceof AdminInterface && 'edit' == $event->getSetting('action')) {
