@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -14,7 +16,7 @@ namespace Sonata\BlockBundle\Exception\Renderer;
 use Sonata\BlockBundle\Model\BlockInterface;
 use Symfony\Component\Debug\Exception\FlattenException;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Templating\EngineInterface;
+use Twig\Environment;
 
 /**
  * This renderer uses a template to display an error message at the block position with extensive debug information.
@@ -23,11 +25,6 @@ use Symfony\Component\Templating\EngineInterface;
  */
 class InlineDebugRenderer implements RendererInterface
 {
-    /**
-     * @var EngineInterface
-     */
-    protected $templating;
-
     /**
      * @var string
      */
@@ -44,14 +41,13 @@ class InlineDebugRenderer implements RendererInterface
     protected $debug;
 
     /**
-     * @param EngineInterface $templating Templating engine
-     * @param string          $template   Template to render
-     * @param bool            $debug      Whether the debug is enabled or not
-     * @param bool            $forceStyle Whether to force style within the template or not
+     * @var Environment
      */
-    public function __construct(EngineInterface $templating, $template, $debug, $forceStyle = true)
+    private $twig;
+
+    public function __construct(Environment $twig, string $template, bool $debug, bool $forceStyle = true)
     {
-        $this->templating = $templating;
+        $this->twig = $twig;
         $this->template = $template;
         $this->debug = $debug;
         $this->forceStyle = $forceStyle;
@@ -60,7 +56,7 @@ class InlineDebugRenderer implements RendererInterface
     /**
      * {@inheritdoc}
      */
-    public function render(\Exception $exception, BlockInterface $block, Response $response = null)
+    public function render(\Exception $exception, BlockInterface $block, Response $response = null): Response
     {
         $response = $response ?: new Response();
 
@@ -82,7 +78,7 @@ class InlineDebugRenderer implements RendererInterface
             'forceStyle' => $this->forceStyle,
         ];
 
-        $content = $this->templating->render($this->template, $parameters);
+        $content = $this->twig->render($this->template, $parameters);
         $response->setContent($content);
 
         return $response;
