@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -83,7 +85,7 @@ class StrategyManager implements StrategyManagerInterface
      *
      * @throws \InvalidArgumentException
      */
-    public function setDefaultFilter($name)
+    public function setDefaultFilter($name): void
     {
         if (!array_key_exists($name, $this->filters)) {
             throw new \InvalidArgumentException(sprintf('Cannot set default exception filter "%s". It does not exist.', $name));
@@ -99,7 +101,7 @@ class StrategyManager implements StrategyManagerInterface
      *
      * @throws \InvalidArgumentException
      */
-    public function setDefaultRenderer($name)
+    public function setDefaultRenderer($name): void
     {
         if (!array_key_exists($name, $this->renderers)) {
             throw new \InvalidArgumentException(sprintf('Cannot set default exception renderer "%s". It does not exist.', $name));
