@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -322,7 +324,7 @@ class BlockHelper extends Helper
      * @param BlockContextInterface $blockContext
      * @param array                 $stats
      */
-    protected function computeAssets(BlockContextInterface $blockContext, array &$stats = null)
+    protected function computeAssets(BlockContextInterface $blockContext, array &$stats = null): void
     {
         if ($blockContext->getBlock()->hasParent()) {
             return;
@@ -397,7 +399,7 @@ class BlockHelper extends Helper
      * @param BlockInterface $block
      * @param array          $stats
      */
-    protected function stopTracing(BlockInterface $block, array $stats)
+    protected function stopTracing(BlockInterface $block, array $stats): void
     {
         $e = $this->traces[$block->getId()]->stop();
 
