@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,7 +18,7 @@ use Sonata\BlockBundle\Block\BlockServiceManager;
 
 class BlockServiceManagerTest extends TestCase
 {
-    public function testGetBlockService()
+    public function testGetBlockService(): void
     {
         $service = $this->createMock('Sonata\BlockBundle\Block\BlockServiceInterface');
 
@@ -33,7 +35,7 @@ class BlockServiceManagerTest extends TestCase
         $this->assertInstanceOf(get_class($service), $manager->get($block));
     }
 
-    public function testInvalidServiceType()
+    public function testInvalidServiceType(): void
     {
         $this->expectException(\RuntimeException::class);
 
@@ -52,7 +54,7 @@ class BlockServiceManagerTest extends TestCase
         $this->assertInstanceOf(get_class($service), $manager->get($block));
     }
 
-    public function testGetBlockServiceException()
+    public function testGetBlockServiceException(): void
     {
         $this->expectException(\RuntimeException::class);
 
@@ -66,7 +68,7 @@ class BlockServiceManagerTest extends TestCase
         $manager->get($block);
     }
 
-    public function testGetEmptyListFromInvalidContext()
+    public function testGetEmptyListFromInvalidContext(): void
     {
         $container = $this->createMock('Symfony\Component\DependencyInjection\ContainerInterface');
         $manager = new BlockServiceManager($container, true);
@@ -78,7 +80,7 @@ class BlockServiceManagerTest extends TestCase
         $this->assertEmpty($manager->getServicesByContext('fake'));
     }
 
-    public function testGetListFromValidContext()
+    public function testGetListFromValidContext(): void
     {
         $container = $this->createMock('Symfony\Component\DependencyInjection\ContainerInterface');
         $manager = new BlockServiceManager($container, true);
@@ -90,7 +92,7 @@ class BlockServiceManagerTest extends TestCase
         $this->assertNotEmpty($manager->getServicesByContext('fake'));
     }
 
-    public function testOrderServices()
+    public function testOrderServices(): void
     {
         $container = $this->createMock('Symfony\Component\DependencyInjection\ContainerInterface');
         $manager = new BlockServiceManager($container, true);
