@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,7 +18,7 @@ use Sonata\BlockBundle\Cache\NoopHttpCacheHandler;
 
 class NoopHttpCacheHandlerTest extends TestCase
 {
-    public function testComputeTtl()
+    public function testComputeTtl(): void
     {
         // check interface
         new NoopHttpCacheHandler();
