@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -17,7 +19,7 @@ use Sonata\BlockBundle\Annotation\Block;
 
 class AnnotationCompilerPassTest extends TestCase
 {
-    public function testMinimalBlock()
+    public function testMinimalBlock(): void
     {
         /*
          * @Block()
@@ -34,7 +36,7 @@ class AnnotationCompilerPassTest extends TestCase
         );
     }
 
-    public function testBlock()
+    public function testBlock(): void
     {
         /*
          * @Block()
