@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -17,7 +19,7 @@ use Sonata\BlockBundle\Event\TextBlockListener;
 
 class TextBlockListenerTest extends TestCase
 {
-    public function testEvent()
+    public function testEvent(): void
     {
         $event = new BlockEvent();
 
@@ -31,7 +33,7 @@ class TextBlockListenerTest extends TestCase
         $this->assertEquals('This block is coming from inline event from the template', $blocks[0]->getSetting('content'));
     }
 
-    public function testEventWithAdmin()
+    public function testEventWithAdmin(): void
     {
         $admin = $this->createMock('Sonata\AdminBundle\Admin\AdminInterface');
         $admin->expects($this->once())->method('getSubject');
