@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -28,7 +30,7 @@ class KeepNoneFilterTest extends TestCase
      *
      * @dataProvider getExceptions
      */
-    public function testFilter(\Exception $exception)
+    public function testFilter(\Exception $exception): void
     {
         // GIVEN
         $block = $this->createMock('Sonata\BlockBundle\Model\BlockInterface');
