@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -24,7 +26,7 @@ class MonkeyThrowRendererTest extends TestCase
     /**
      * test the render() method with a standard Exception.
      */
-    public function testRenderWithStandardException()
+    public function testRenderWithStandardException(): void
     {
         $this->expectException(\Exception::class);
 
@@ -43,7 +45,7 @@ class MonkeyThrowRendererTest extends TestCase
     /**
      * test the render() method with another exception to ensure it correctly throws the provided exception.
      */
-    public function testRenderWithRuntimeException()
+    public function testRenderWithRuntimeException(): void
     {
         $this->expectException(\RuntimeException::class);
 
