@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,6 +18,7 @@ use Sonata\BlockBundle\Exception\Filter\FilterInterface;
 use Sonata\BlockBundle\Exception\Renderer\RendererInterface;
 use Sonata\BlockBundle\Exception\Strategy\StrategyManager;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  * Test the Exception Strategy Manager.
@@ -77,7 +80,7 @@ class StrategyManagerTest extends TestCase
     /**
      * setup a basic scenario to avoid long test setup.
      */
-    public function setUp()
+    public function setUp(): void
     {
         $this->renderer1 = $this->createMock('\Sonata\BlockBundle\Exception\Renderer\RendererInterface');
         $this->renderer2 = $this->createMock('\Sonata\BlockBundle\Exception\Renderer\RendererInterface');
@@ -117,7 +120,7 @@ class StrategyManagerTest extends TestCase
     /**
      * test getBlockRenderer() with existing block renderer.
      */
-    public function testGetBlockRendererWithExisting()
+    public function testGetBlockRendererWithExisting(): void
     {
         // GIVEN
         $block = $this->getMockBlock('block.type1');
@@ -133,7 +136,7 @@ class StrategyManagerTest extends TestCase
     /**
      * test getBlockRenderer() with non existing block renderer.
      */
-    public function testGetBlockRendererWithNonExisting()
+    public function testGetBlockRendererWithNonExisting(): void
     {
         // GIVEN
         $block = $this->getMockBlock('block.other_type');
@@ -149,7 +152,7 @@ class StrategyManagerTest extends TestCase
     /**
      * test getBlockFilter() with an existing block filter.
      */
-    public function testGetBlockFilterWithExisting()
+    public function testGetBlockFilterWithExisting(): void
     {
         // GIVEN
         $block = $this->getMockBlock('block.type1');
@@ -165,7 +168,7 @@ class StrategyManagerTest extends TestCase
     /**
      * test getting the default block renderer.
      */
-    public function testGetBlockFilterWithNonExisting()
+    public function testGetBlockFilterWithNonExisting(): void
     {
         // GIVEN
         $block = $this->getMockBlock('block.other_type');
@@ -181,7 +184,7 @@ class StrategyManagerTest extends TestCase
     /**
      * test handleException() with a keep none filter.
      */
-    public function testHandleExceptionWithKeepNoneFilter()
+    public function testHandleExceptionWithKeepNoneFilter(): void
     {
         // GIVEN
         $this->filter1->expects($this->once())->method('handle')->will($this->returnValue(false));
@@ -201,11 +204,13 @@ class StrategyManagerTest extends TestCase
     /**
      * test handleException() with a keep all filter.
      */
-    public function testHandleExceptionWithKeepAllFilter()
+    public function testHandleExceptionWithKeepAllFilter(): void
     {
+        $rendererResponse = new Response();
+        $rendererResponse->setContent('renderer response');
         // GIVEN
         $this->filter1->expects($this->once())->method('handle')->will($this->returnValue(true));
-        $this->renderer1->expects($this->once())->method('render')->will($this->returnValue('renderer response'));
+        $this->renderer1->expects($this->once())->method('render')->will($this->returnValue($rendererResponse));
 
         $exception = new \Exception();
         $block = $this->getMockBlock('block.other_type');
@@ -215,7 +220,7 @@ class StrategyManagerTest extends TestCase
 
         // THEN
         $this->assertNotNull($response, 'should return something');
-        $this->assertEquals('renderer response', $response, 'should return the renderer response');
+        $this->assertEquals('renderer response', $response->getContent(), 'should return the renderer response');
     }
 
     /**
