@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -21,7 +23,7 @@ use Symfony\Component\HttpFoundation\Response;
 
 class BlockHelperTest extends TestCase
 {
-    public function testRenderEventWithNoListener()
+    public function testRenderEventWithNoListener(): void
     {
         $blockServiceManager = $this->createMock('Sonata\BlockBundle\Block\BlockServiceManagerInterface');
         $blockRenderer = $this->createMock('Sonata\BlockBundle\Block\BlockRendererInterface');
@@ -36,7 +38,7 @@ class BlockHelperTest extends TestCase
         $this->assertEquals('', $helper->renderEvent('my.event'));
     }
 
-    public function testRenderEventWithListeners()
+    public function testRenderEventWithListeners(): void
     {
         $blockService = $this->createMock('Sonata\BlockBundle\Block\BlockServiceInterface');
         $blockService->expects($this->once())->method('getJavascripts')->will($this->returnValue([
