@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -31,7 +33,7 @@ class BlockExtensionTest extends TestCase
      */
     protected $env;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->blockHelper = $this->getMockBuilder(
             'Sonata\BlockBundle\Templating\Helper\BlockHelper'
@@ -66,7 +68,7 @@ class BlockExtensionTest extends TestCase
     /**
      * @dataProvider provideFunction
      */
-    public function testFunction($name, $args, $expectedMethod)
+    public function testFunction($name, $args, $expectedMethod): void
     {
         $this->blockHelper->expects($this->once())
             ->method($expectedMethod);
