@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,7 +18,7 @@ use Sonata\BlockBundle\Twig\GlobalVariables;
 
 class GlobalVariablesTest extends TestCase
 {
-    public function testGlobalVariables()
+    public function testGlobalVariables(): void
     {
         $variables = new GlobalVariables([]);
 
