@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Sonata Project package.
  *
@@ -16,7 +18,7 @@ use Sonata\BlockBundle\Util\RecursiveBlockIteratorIterator;
 
 class RecursiveBlockIteratorIteratorTest extends TestCase
 {
-    public function testInterface()
+    public function testInterface(): void
     {
         $block2 = $this->createMock('Sonata\BlockBundle\Model\BlockInterface');
         $block2->expects($this->any())->method('getType')->will($this->returnValue('block2'));
